requirements.txt
langchain
langchain-community
langchain-groq
pypdf
tiktoken
python-dotenv
faiss-cpu
sentence-transformers
streamlit
